feat: responsive layout and sidebar improvements
- Implemented responsive behavior for the sidebar:
  - Sidebar width fixed to 250px on large screens.
  - Sidebar height dynamically adjusted to fit the screen (between header and footer).
  - Sidebar collapses smoothly on small screens with transition animations.
  - Toggle button added for opening/closing the sidebar on smaller screens.

- Adjusted main container layout:
  - Ensured the content fills the available space when the sidebar is collapsed.
  - Added calc(100vh - headerHeight - footerHeight) logic for proper height calculation.
  - Fixed the issue where the sidebar space was being occupied even when hidden.

- Prepared CSS variables for dark and light mode themes:
  - Variables for --primary-color, --secondary-color, --text-color, and --border-color.
  - Applied these variables consistently to header, footer, sidebar, and content.

- Modularized CSS and components:
  - Header, Sidebar, Footer, and Layout CSS split into separate modules for better maintainability.

- Minor visual adjustments:
  - Adjusted box shadows, padding, and flex alignment for better aesthetics.
  - Fixed spacing and alignment issues across layouts in both light and dark modes.
