+++ src/components/Layout/Footer/Footer.module.css
@@ -1,13 +1,11 @@
 .footer {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-  height: 40px;
-  background-color: #f8f8f8;
-  color: #333;
+  height: 40px; /* Altura fixa */
+  background-color: var(--header-footer-color); /* Fundo adaptável */
   display: flex;
   align-items: center;
   justify-content: center;
-  box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
-  z-index: 1000;
+  text-align: center;
+  font-size: 0.9rem;
+  box-shadow: 0 -2px 4px var(--border-color); /* Sombras variáveis */
+  color: var(--text-color); /* Texto adaptável */
 }
