+++ src/components/Layout/Header/Header.module.css
@@ -1,14 +1,24 @@
 .header {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 60px;
-  background-color: #f8f8f8;
-  color: #333;
+  height: 60px; /* Altura fixa */
+  background-color: var(--header-footer-color); /* Cor adaptada para dark/light */
   display: flex;
   align-items: center;
-  justify-content: center;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  justify-content: space-between; /* Espaçamento entre elementos */
+  padding: 0 1rem;
+  box-shadow: 0 2px 4px var(--border-color); /* Sombras variáveis */
   z-index: 1000;
+}
+
+.menuButton {
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  color: var(--text-color); /* Texto adaptável */
+  cursor: pointer;
+}
+
+@media (min-width: 768px) {
+  .menuButton {
+    display: none; /* Esconde botão no desktop */
+  }
 }
