+++ src/components/Layout/Layout.module.css
@@ -1,48 +1,24 @@
-/* Layout.module.css */
 .layout {
-  display: flex; /* Organiza Sidebar e MainContent lado a lado */
-  min-height: 100vh; /* Garante que a altura total do layout seja sempre 100% da viewport */
-  flex-direction: column; /* Organização vertical para Header, Main e Footer */
-}
-
-.header {
-  height: 60px; /* Defina a altura fixa do Header */
-  background-color: var(--primary-color);
   display: flex;
-  align-items: center;
-  padding: 0 1rem;
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-  z-index: 100;
+  flex-direction: column; /* Organização vertical */
+  min-height: 100vh; /* Altura total da tela */
+  background-color: var(--primary-color); /* Fundo adaptável */
+  color: var(--text-color); /* Texto adaptável */
 }
 
 .mainContainer {
-  display: flex; /* Sidebar e conteúdo principal lado a lado */
-  flex: 1; /* Ocupar todo o espaço restante entre Header e Footer */
-  overflow: hidden; /* Impede o overflow */
+  display: flex;
+  flex: 1;
+  overflow: hidden;
+  position: relative;
+  min-height: calc(100vh - 100px); /* Ajusta para Header (60px) e Footer (40px) */
 }
 
-.sidebar {
-  width: 250px; /* Largura fixa da Sidebar */
-  background-color: var(--secondary-color);
-  display: flex;
-  flex-direction: column;
-  padding: 1rem;
-  box-shadow: 2px 0 4px rgba(0, 0, 0, 0.1);
-}
 
 .content {
-  flex: 1; /* Garante que o conteúdo principal ocupe o espaço restante */
+  flex: 1; /* Ocupa espaço restante */
   padding: 1rem;
-  overflow-y: auto; /* Permite rolagem apenas no conteúdo */
-}
-
-.footer {
-  height: 40px; /* Altura fixa */
-  background-color: var(--primary-color);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  text-align: center;
-  font-size: 0.9rem;
-  box-shadow: 0 -2px 4px rgba(0, 0, 0, 0.1);
+  overflow-y: auto; /* Rola apenas o conteúdo */
+  background-color: var(--primary-color); /* Fundo adaptável */
+  color: var(--text-color); /* Texto adaptável */
 }
