+++ src/components/Layout/Layout.tsx
@@ -1,5 +1,7 @@
-// Layout.tsx
-import React from "react";
+import React, { useState } from "react";
+import Header from "@components/Layout/Header/Header";
+import Sidebar from "@components/Layout/Sidebar/Sidebar";
+import Footer from "@components/Layout/Footer/Footer";
 import styles from "./Layout.module.css";
 
 interface LayoutProps {
@@ -7,34 +9,20 @@ interface LayoutProps {
 }
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
+  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+
+  const toggleSidebar = () => {
+    setIsSidebarOpen(!isSidebarOpen);
+  };
+
   return (
     <div className={styles.layout}>
-      {/* Header */}
-      <header className={styles.header}>
-        <h1>Task Manager Chat</h1>
-      </header>
-
-      {/* Main Section */}
+      <Header toggleSidebar={toggleSidebar} />
       <div className={styles.mainContainer}>
-        {/* Sidebar */}
-        <aside className={styles.sidebar}>
-          <nav>
-            <ul>
-              <li><a href="/dashboard">Dashboard</a></li>
-              <li><a href="/tasks">Tasks</a></li>
-              <li><a href="/settings">Settings</a></li>
-            </ul>
-          </nav>
-        </aside>
-
-        {/* Main Content */}
+        <Sidebar isOpen={isSidebarOpen} toggleSidebar={toggleSidebar} />
         <main className={styles.content}>{children}</main>
       </div>
-
-      {/* Footer */}
-      <footer className={styles.footer}>
-        © 2024 Task Manager Chat. All rights reserved.
-      </footer>
+      <Footer />
     </div>
   );
 };
