+++ src/index.tsx
@@ -1,10 +1,11 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import Layout from '@components/Layout/Layout';
-import Home from '@pages/Home/Home'; 
+import React from "react";
+import ReactDOM from "react-dom/client";
+import Layout from "@components/Layout/Layout";
+import Home from "@pages/Home/Home";
+import "@styles/index.css";
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+  document.getElementById("root") as HTMLElement
 );
 
 root.render(
